one more reshuffle, Modifier is it's own class to be used by Policy
