+++ tests/test_5_policies.py
@@ -3,7 +3,7 @@ from pam.policy import policies
 from pam.policy import modifiers
 from pam.policy import filters
 from pam.policy import probability_samplers
-from pam.policy import policy_levels
+from pam.policy import policies
 from tests.fixtures import *
 
 
@@ -161,7 +161,7 @@ def test_home_education_home_removal_of_education_act(person_home_education_home
     household = instantiate_household_with([person_home_education_home])
     assert_correct_activities(person=household.people['1'], ordered_activities_list=['home', 'education', 'home'])
 
-    policy = policy_levels.ActivityPolicy(modifiers.RemoveActivity(activities=['education']), 1)
+    policy = policies.ActivityPolicy(modifiers.RemoveActivity(activities=['education']), 1)
     policy.apply_to(household)
     assert_correct_activities(person=household.people['1'], ordered_activities_list=['home'])
 
@@ -181,7 +181,7 @@ def test_home_education_home_education_home_removal_of_education_act():
     assert_correct_activities(person=household.people['1'],
                               ordered_activities_list=['home', 'education', 'home', 'education', 'home'])
 
-    policy = policy_levels.ActivityPolicy(modifiers.RemoveActivity(activities=['education']), 1)
+    policy = policies.ActivityPolicy(modifiers.RemoveActivity(activities=['education']), 1)
     policy.apply_to(household)
     assert_correct_activities(person=household.people['1'], ordered_activities_list=['home'])
 
@@ -276,7 +276,7 @@ def test_home_work_home_education_home_removal_of_education_act():
     household = instantiate_household_with([person])
     assert_correct_activities(person=household.people['1'], ordered_activities_list=['home', 'work', 'home', 'education', 'home'])
 
-    policy = policy_levels.ActivityPolicy(modifiers.RemoveActivity(activities=['education']), 1)
+    policy = policies.ActivityPolicy(modifiers.RemoveActivity(activities=['education']), 1)
     policy.apply_to(household)
     assert_correct_activities(person=household.people['1'], ordered_activities_list=['home', 'work', 'home'])
 
@@ -295,7 +295,7 @@ def test_attribute_based_remove_activity_policy_removes_all_matching_activities_
     assert age_condition_over_17(household.people['1'].attributes['age'])
     assert job_condition_education(household.people['1'].attributes['job'])
 
-    policy_remove_higher_education = policy_levels.ActivityPolicy(
+    policy_remove_higher_education = policies.ActivityPolicy(
         modifiers.RemoveActivity(['education']),
         probability_samplers.ActivityProbability(['education'], 1),
         filters.PersonAttributeFilter(conditions={'age': age_condition_over_17, 'job': job_condition_education}, how='all')
@@ -319,7 +319,7 @@ def test_attribute_based_remove_activity_policy_does_not_remove_matching_activit
     assert age_condition_over_17(household.people['1'].attributes['age'])
     assert not job_condition_wasevrrr(household.people['1'].attributes['job'])
 
-    policy_remove_higher_education = policy_levels.ActivityPolicy(
+    policy_remove_higher_education = policies.ActivityPolicy(
         modifiers.RemoveActivity(['education']),
         probability_samplers.ActivityProbability(['education'], 1),
         filters.PersonAttributeFilter(conditions={'age': age_condition_over_17, 'job': job_condition_wasevrrr},
@@ -345,7 +345,7 @@ def test_attribute_based_remove_activity_policy_removes_all_matching_activities_
     assert age_condition_over_17(household.people['1'].attributes['age'])
     assert not job_condition_wasevrrr(household.people['1'].attributes['job'])
 
-    policy_remove_higher_education = policy_levels.ActivityPolicy(
+    policy_remove_higher_education = policies.ActivityPolicy(
         modifiers.RemoveActivity(['education']),
         probability_samplers.ActivityProbability(['education'], 1),
         filters.PersonAttributeFilter(conditions={'age': age_condition_over_17, 'job': job_condition_wasevrrr},
@@ -371,7 +371,7 @@ def test_attribute_based_remove_activity_policy_does_not_remove_matching_activit
     assert not age_condition_under_0(household.people['1'].attributes['age'])
     assert not job_condition_wasevrrr(household.people['1'].attributes['job'])
 
-    policy_remove_higher_education = policy_levels.ActivityPolicy(
+    policy_remove_higher_education = policies.ActivityPolicy(
         modifiers.RemoveActivity(['education']),
         probability_samplers.ActivityProbability(['education'], 1),
         filters.PersonAttributeFilter(conditions={'age': age_condition_under_0, 'job': job_condition_wasevrrr},
@@ -408,7 +408,7 @@ def test_evaluate_activity_policy_selects_steve_for_individual_activity_removal(
     assert_correct_activities(person=bobby, ordered_activities_list=['home', 'education', 'home'])
 
     # i.e. First of Steve's work activities is affected and only that activity is affected
-    policy = policy_levels.ActivityPolicy(
+    policy = policies.ActivityPolicy(
         modifiers.RemoveActivity(['education', 'escort', 'leisure', 'shop', 'work']),
         probability_samplers.ActivityProbability(['education', 'escort', 'leisure', 'shop', 'work'], 0.5)
     )
@@ -425,7 +425,7 @@ def test_household_policy_with_household_based_probability(SmithHousehold, mocke
     mocker.patch.object(random, 'random', side_effect=[0])
     household = SmithHousehold
     # i.e. household is affected and affects activities on household level
-    policy = policy_levels.HouseholdPolicy(
+    policy = policies.HouseholdPolicy(
         modifiers.RemoveActivity(['education', 'escort', 'leisure', 'shop', 'work']),
         probability_samplers.HouseholdProbability(0.5))
     policy.apply_to(household)
@@ -458,7 +458,7 @@ def test_household_policy_with_household_based_probability_with_a_satisfied_pers
         people_satisfying_age_condition_under_10 += discrete_sampler(person, age_mapping, age_distribution)
     assert people_satisfying_age_condition_under_10 == 1
 
-    policy = policy_levels.HouseholdPolicy(
+    policy = policies.HouseholdPolicy(
         modifiers.RemoveActivity(['education', 'escort', 'leisure', 'shop', 'work']),
         [probability_samplers.HouseholdProbability(0.5),
          probability_samplers.PersonProbability(discrete_sampler, {'mapping': age_mapping, 'distribution': age_distribution})]
@@ -473,7 +473,7 @@ def test_household_policy_with_person_based_probability(SmithHousehold, mocker):
     mocker.patch.object(random, 'random', side_effect=[0.06249])
     household = SmithHousehold
     # i.e. Bobby is affected and affects activities on household level
-    policy = policy_levels.HouseholdPolicy(
+    policy = policies.HouseholdPolicy(
         modifiers.RemoveActivity(['education', 'escort', 'leisure', 'shop', 'work']),
         probability_samplers.PersonProbability(0.5))
     policy.apply_to(household)
@@ -506,7 +506,7 @@ def test_household_policy_with_person_based_probability_with_a_satisfied_person_
         people_satisfying_age_condition_under_10 += discrete_sampler(person, age_mapping, age_distribution)
     assert people_satisfying_age_condition_under_10 == 1
 
-    policy = policy_levels.HouseholdPolicy(
+    policy = policies.HouseholdPolicy(
         modifiers.RemoveActivity(['education', 'escort', 'leisure', 'shop', 'work']),
         probability_samplers.PersonProbability(discrete_sampler, {'mapping': age_mapping, 'distribution': age_distribution})
     )
@@ -520,7 +520,7 @@ def test_household_policy_with_activity_based_probability(SmithHousehold, mocker
     mocker.patch.object(random, 'random', side_effect=[0.000244140624])
     household = SmithHousehold
     # i.e. Bobby's education activity is affected and affects activities on household level
-    policy = policy_levels.HouseholdPolicy(
+    policy = policies.HouseholdPolicy(
         modifiers.RemoveActivity(['education', 'escort', 'leisure', 'shop', 'work']),
         probability_samplers.ActivityProbability(['education', 'escort', 'leisure', 'shop', 'work'], 0.5))
     policy.apply_to(household)
@@ -553,7 +553,7 @@ def test_household_policy_with_activity_based_probability_with_a_satisfied_perso
         people_satisfying_age_condition_under_10 += discrete_sampler(person, age_mapping, age_distribution)
     assert people_satisfying_age_condition_under_10 == 1
 
-    policy = policy_levels.HouseholdPolicy(
+    policy = policies.HouseholdPolicy(
         modifiers.RemoveActivity(['education', 'escort', 'leisure', 'shop', 'work']),
         [probability_samplers.ActivityProbability(['education', 'escort', 'leisure', 'shop', 'work'], 0.5),
          probability_samplers.PersonProbability(discrete_sampler, {'mapping': age_mapping, 'distribution': age_distribution})]
@@ -568,7 +568,7 @@ def test_person_policy_with_person_based_probability(mocker, SmithHousehold):
     mocker.patch.object(random, 'random', side_effect=[1, 1, 1, 0])
     household = SmithHousehold
     # i.e. Bobby is affected and his activities are the only one affected in household
-    policy = policy_levels.PersonPolicy(
+    policy = policies.PersonPolicy(
         modifiers.RemoveActivity(['education', 'escort', 'leisure', 'shop', 'work']),
         probability_samplers.PersonProbability(0.5))
     bobby = household.people['4']
@@ -602,7 +602,7 @@ def test_person_policy_with_person_based_probability_with_a_satisfied_person_att
         people_satisfying_age_condition_under_10 += discrete_sampler(person, age_mapping, age_distribution)
     assert people_satisfying_age_condition_under_10 == 1
 
-    policy = policy_levels.PersonPolicy(
+    policy = policies.PersonPolicy(
         modifiers.RemoveActivity(['education', 'escort', 'leisure', 'shop', 'work']),
         probability_samplers.PersonProbability(discrete_sampler, {'mapping': age_mapping, 'distribution': age_distribution}))
     bobby = household.people['4']
@@ -616,7 +616,7 @@ def test_person_policy_with_activity_based_probability(SmithHousehold, mocker):
     mocker.patch.object(random, 'random', side_effect=[0] + [1] * 11)
     household = SmithHousehold
     # i.e. First of Steve's work activities is affected and affects all listed activities for just Steve
-    policy = policy_levels.PersonPolicy(
+    policy = policies.PersonPolicy(
         modifiers.RemoveActivity(['education', 'escort', 'leisure', 'shop', 'work']),
         probability_samplers.ActivityProbability(['education', 'escort', 'leisure', 'shop', 'work'], 0.5)
     )
@@ -651,7 +651,7 @@ def test_person_policy_with_activity_based_probability_with_a_satisfied_person_a
         people_satisfying_age_condition_over_20 += discrete_sampler(person, age_mapping, age_distribution)
     assert people_satisfying_age_condition_over_20 == 2
 
-    policy = policy_levels.PersonPolicy(
+    policy = policies.PersonPolicy(
         modifiers.RemoveActivity(['education', 'escort', 'leisure', 'shop', 'work']),
         [probability_samplers.ActivityProbability(['education', 'escort', 'leisure', 'shop', 'work'], 0.5),
          probability_samplers.PersonProbability(discrete_sampler, {'mapping': age_mapping, 'distribution': age_distribution})]
@@ -674,7 +674,7 @@ def test_MoveActivityToHomeLocation_moves_shop_to_home_location():
     Hilda.add(Activity(5, 'home', 'a', start_time=mtdt(17 * 60), end_time=END_OF_DAY))
     hhld = instantiate_household_with([Hilda])
 
-    policy = policy_levels.PersonPolicy(
+    policy = policies.PersonPolicy(
         modifiers.MoveActivityTourToHomeLocation(['shop']),
         probability_samplers.PersonProbability(1)
     )
@@ -694,7 +694,7 @@ def test_MoveActivityToHomeLocation_updates_Legs_after_moving_shopping_trip():
 
     hhld = instantiate_household_with([Hilda])
 
-    policy = policy_levels.PersonPolicy(
+    policy = policies.PersonPolicy(
         modifiers.MoveActivityTourToHomeLocation(['shop']),
         probability_samplers.PersonProbability(1)
     )
@@ -713,7 +713,7 @@ def test_MoveActivityToHomeLocation_performs_mode_shift_after_moving_shopping_tr
     Hilda.add(Activity(5, 'home', 'a', start_time=mtdt(17 * 60), end_time=END_OF_DAY))
     hhld = instantiate_household_with([Hilda])
 
-    policy = policy_levels.PersonPolicy(
+    policy = policies.PersonPolicy(
         modifiers.MoveActivityTourToHomeLocation(['shop']),
         probability_samplers.PersonProbability(1)
     )
@@ -732,7 +732,7 @@ def test_MoveActivityToHomeLocation_performs_mode_shift_to_car_due_to_driving_li
     Hilda.add(Activity(5, 'home', 'a', start_time=mtdt(17 * 60), end_time=END_OF_DAY))
     hhld = instantiate_household_with([Hilda])
 
-    policy = policy_levels.PersonPolicy(
+    policy = policies.PersonPolicy(
         modifiers.MoveActivityTourToHomeLocation(['shop']),
         probability_samplers.PersonProbability(1)
     )
@@ -752,7 +752,7 @@ def test_MoveActivityToHomeLocation_performs_mode_shift_to_walk_due_to_lack_of_d
     Hilda.add(Activity(5, 'home', 'a', start_time=mtdt(17 * 60), end_time=END_OF_DAY))
     hhld = instantiate_household_with([Hilda])
 
-    policy = policy_levels.PersonPolicy(
+    policy = policies.PersonPolicy(
         modifiers.MoveActivityTourToHomeLocation(['shop']),
         probability_samplers.PersonProbability(1)
     )
@@ -788,7 +788,7 @@ def test_MoveActivityToHomeLocation_moves_shopping_tour_to_home_location(SmithHo
     assert_correct_activities(person=Bobby, ordered_activities_list=['home', 'education', 'home'])
     assert_correct_activities_locations(person=Bobby, ordered_activities_locations_list=['a', 'b', 'a'])
 
-    policy = policy_levels.PersonPolicy(
+    policy = policies.PersonPolicy(
         modifiers.MoveActivityTourToHomeLocation(['shop_1', 'shop_2']),
         probability_samplers.PersonProbability(1)
     )
@@ -821,7 +821,7 @@ def test_MoveActivityToHomeLocation_moves_partial_fit_shopping_tours(SmithHouseh
     assert_correct_activities(person=Hilda, ordered_activities_list=['home', 'shop_1', 'home', 'shop_2', 'home'])
     assert_correct_activities_locations(person=Hilda, ordered_activities_locations_list=['a', 'b', 'a', 'c', 'a'])
 
-    policy = policy_levels.PersonPolicy(
+    policy = policies.PersonPolicy(
         modifiers.MoveActivityTourToHomeLocation(['shop_1', 'shop_2']),
         probability_samplers.PersonProbability(1)
     )
@@ -850,7 +850,7 @@ def test_MoveActivityToHomeLocation_does_moves_only_valid_shopping_tour(SmithHou
     assert_correct_activities(person=Hilda, ordered_activities_list=['home', 'shop', 'home', 'shop', 'leisure', 'home'])
     assert_correct_activities_locations(person=Hilda, ordered_activities_locations_list=['a', 'b', 'a', 'c', 'c', 'a'])
 
-    policy = policy_levels.PersonPolicy(
+    policy = policies.PersonPolicy(
         modifiers.MoveActivityTourToHomeLocation(['shop']),
         probability_samplers.PersonProbability(1)
     )
