+++ tests/test_5_policy_levels.py
@@ -1,6 +1,6 @@
 from pam.policy import modifiers
 from pam.policy import probability_samplers
-from pam.policy import policy_levels
+from pam.policy import policies
 from tests.fixtures import *
 
 
@@ -15,7 +15,7 @@ def assert_correct_activities(person, ordered_activities_list):
 
 def test_HouseholdPolicy_verifies_for_appropriate_probabilities(mocker):
     mocker.patch.object(probability_samplers, 'verify_probability')
-    policy_levels.HouseholdPolicy(modifiers.RemoveActivity(['']), 0.5)
+    policies.HouseholdPolicy(modifiers.RemoveActivity(['']), 0.5)
 
     probability_samplers.verify_probability.assert_called_once_with(0.5)
 
@@ -24,7 +24,7 @@ def test_HouseholdPolicy_apply_to_delegates_to_modifier_policy_apply_to_for_sing
     mocker.patch.object(modifiers.RemoveActivity, 'apply_to')
     mocker.patch.object(probability_samplers.SimpleProbability, 'sample', return_value=True)
 
-    policy = policy_levels.HouseholdPolicy(modifiers.RemoveActivity(['']), probability_samplers.SimpleProbability(1.))
+    policy = policies.HouseholdPolicy(modifiers.RemoveActivity(['']), probability_samplers.SimpleProbability(1.))
     household = SmithHousehold
 
     policy.apply_to(household)
@@ -36,7 +36,7 @@ def test_HouseholdPolicy_apply_to_delegates_to_modifier_policy_apply_to_for_list
     mocker.patch.object(modifiers.RemoveActivity, 'apply_to')
     mocker.patch.object(probability_samplers.SimpleProbability, 'p', return_value=1)
 
-    policy = policy_levels.HouseholdPolicy(modifiers.RemoveActivity(['']), [1., probability_samplers.SimpleProbability(1.)])
+    policy = policies.HouseholdPolicy(modifiers.RemoveActivity(['']), [1., probability_samplers.SimpleProbability(1.)])
     household = SmithHousehold
 
     policy.apply_to(household)
@@ -46,7 +46,7 @@ def test_HouseholdPolicy_apply_to_delegates_to_modifier_policy_apply_to_for_list
 
 def test_PersonPolicy_verifies_for_appropriate_probabilities(mocker):
     mocker.patch.object(probability_samplers, 'verify_probability')
-    policy_levels.PersonPolicy(modifiers.RemoveActivity(['']), 0.5)
+    policies.PersonPolicy(modifiers.RemoveActivity(['']), 0.5)
 
     probability_samplers.verify_probability.assert_called_once_with(0.5, (probability_samplers.HouseholdProbability))
 
@@ -55,7 +55,7 @@ def test_PersonPolicy_apply_to_delegates_to_modifier_policy_apply_to_for_single_
     mocker.patch.object(modifiers.RemoveActivity, 'apply_to')
     mocker.patch.object(probability_samplers.SimpleProbability, 'sample', return_value=True)
 
-    policy = policy_levels.PersonPolicy(modifiers.RemoveActivity(['']), probability_samplers.SimpleProbability(1.))
+    policy = policies.PersonPolicy(modifiers.RemoveActivity(['']), probability_samplers.SimpleProbability(1.))
     household = SmithHousehold
 
     policy.apply_to(household)
@@ -67,7 +67,7 @@ def test_PersonPolicy_apply_to_delegates_to_modifier_policy_apply_to_for_list_of
     mocker.patch.object(modifiers.RemoveActivity, 'apply_to')
     mocker.patch.object(probability_samplers.SimpleProbability, 'p', return_value=1)
 
-    policy = policy_levels.PersonPolicy(modifiers.RemoveActivity(['']), [1., probability_samplers.SimpleProbability(1.)])
+    policy = policies.PersonPolicy(modifiers.RemoveActivity(['']), [1., probability_samplers.SimpleProbability(1.)])
     household = SmithHousehold
 
     policy.apply_to(household)
@@ -77,7 +77,7 @@ def test_PersonPolicy_apply_to_delegates_to_modifier_policy_apply_to_for_list_of
 
 def test_ActivityPolicy_verifies_for_appropriate_probabilities(mocker):
     mocker.patch.object(probability_samplers, 'verify_probability')
-    policy_levels.ActivityPolicy(modifiers.RemoveActivity(['']), 0.5)
+    policies.ActivityPolicy(modifiers.RemoveActivity(['']), 0.5)
 
     probability_samplers.verify_probability.assert_called_once_with(
         0.5, (probability_samplers.HouseholdProbability, probability_samplers.PersonProbability))
@@ -87,7 +87,7 @@ def test_ActivityPolicy_apply_to_delegates_to_modifier_policy_apply_to_for_singl
     mocker.patch.object(modifiers.RemoveActivity, 'apply_to')
     mocker.patch.object(probability_samplers.SimpleProbability, 'sample', return_value=True)
 
-    policy = policy_levels.ActivityPolicy(modifiers.RemoveActivity(['']), probability_samplers.SimpleProbability(1.))
+    policy = policies.ActivityPolicy(modifiers.RemoveActivity(['']), probability_samplers.SimpleProbability(1.))
     household = SmithHousehold
 
     policy.apply_to(household)
@@ -99,7 +99,7 @@ def test_ActivityPolicy_apply_to_delegates_to_modifier_policy_apply_to_for_list_
     mocker.patch.object(modifiers.RemoveActivity, 'apply_to')
     mocker.patch.object(probability_samplers.SimpleProbability, 'p', return_value=1)
 
-    policy = policy_levels.ActivityPolicy(modifiers.RemoveActivity(['']), [1., probability_samplers.SimpleProbability(1.)])
+    policy = policies.ActivityPolicy(modifiers.RemoveActivity(['']), [1., probability_samplers.SimpleProbability(1.)])
     household = SmithHousehold
 
     policy.apply_to(household)
